자기 정보만 수정삭제가능

+++ src/main/java/com/example/prj1/member/service/MemberService.java
@@ -61,34 +61,42 @@ public class MemberService {
 
     }
 
-    public boolean remove(MemberForm data) {
-        Member member = memberRepository.findById(data.getId()).get();
-        String dbPw = member.getPassword();
-        String formPw = data.getPassword();
-        if (dbPw.equals(formPw)) {
-            memberRepository.delete(member);
-            return true;
-        } else {
-            return false;
+    public boolean remove(MemberForm data, MemberDto user) {
+        if (user != null) {
+
+            Member member = memberRepository.findById(data.getId()).get();
+            if (member.getId().equals(user.getId())) {
+                String dbPw = member.getPassword();
+                String formPw = data.getPassword();
+                if (dbPw.equals(formPw)) {
+                    memberRepository.delete(member);
+                    return true;
+                }
+            }
         }
+        return false;
     }
 
-    public boolean update(MemberForm data) {
-        //조회
-        Member member = memberRepository.findById(data.getId()).get();
-        String dbPw = member.getPassword();
-        String formPw = data.getPassword();
-        if (dbPw.equals(formPw)) {
-            //변경
-            member.setNickName(data.getNickName());
-            member.setInfo(data.getInfo());
-            //저장
-            memberRepository.save(member);
+    public boolean update(MemberForm data, MemberDto user) {
+        if (user != null) {
+            //조회
+            Member member = memberRepository.findById(data.getId()).get();
+            if (member.getId().equals(user.getId())) {
 
-            return true;
-        } else {
-            return false;
+                String dbPw = member.getPassword();
+                String formPw = data.getPassword();
+                if (dbPw.equals(formPw)) {
+                    //변경
+                    member.setNickName(data.getNickName());
+                    member.setInfo(data.getInfo());
+                    //저장
+                    memberRepository.save(member);
+
+                    return true;
+                }
+            }
         }
+        return false;
     }
 
     public boolean updatePassword(String id, String oldPassword, String newPassword) {
